Adding enable config option to opt in/out and automatically disable notification server for cloud providers

+++ lib/vagrant-notify/action/stop_server.rb
@@ -7,6 +7,9 @@ module Vagrant
         end
 
         def call(env)
+          
+          return if env[:machine].config.notify.enable == false
+
           if env[:machine].state.id == :running
             env[:machine].communicate.sudo('rm /usr/bin/notify-send; exit 0')
             env[:machine].communicate.sudo('mv /usr/bin/{notify-send.bkp,notify-send}; exit 0')
